Rename package and module files.
