+++ .gitignore
@@ -191,6 +191,9 @@ Thumbs.db
 *.tar.gz
 *.rar
 
+# Local Files
+local.properties
+
 # virtual machine crash logs (Reference: http://www.java.com/en/download/help/error_hotspot.xml)
 hs_err_pid*
