+++ .idea/modules.xml
@@ -2,8 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/RandomNumRoller.iml" filepath="$PROJECT_DIR$/.idea/modules/RandomNumRoller.iml" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/app/RandomNumRoller.app.iml" filepath="$PROJECT_DIR$/.idea/modules/app/RandomNumRoller.app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/RothGar.iml" filepath="$PROJECT_DIR$/.idea/modules/RothGar.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/app/RothGar.app.iml" filepath="$PROJECT_DIR$/.idea/modules/app/RothGar.app.iml" />
     </modules>
   </component>
 </project>
