+++ app/build.gradle
@@ -9,7 +9,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "com.projectmalthorn.randomnumroller"
+        applicationId "com.glenjarvis.rothgar"
         minSdkVersion 26
         targetSdkVersion 30
         versionCode 1
