+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
 <!--        <data>-->
 <!--            <variable-->
 <!--                name="mainActivity"-->
-<!--                type="com.projectmalthorn.randomnumroller.MainActivity" />-->
+<!--                type="rothgarivity" />-->
 <!--        </data>-->
 
     <androidx.constraintlayout.widget.ConstraintLayout
