+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
-    <string name="app_name">RandomNumRoller</string>
+    <string name="app_name">RothGar</string>
     <string name="roll">Roll!</string>
 </resources>
