+++ settings.gradle
@@ -1,2 +1,2 @@
 include ':app'
-rootProject.name = "RandomNumRoller"
+rootProject.name = "RothGar"
